Initial work on Japanese syllabel splitting
Not all of the tests are passing, but I need to change direction because
this is not the right approach. Japanese does not have syllables, but
the sino-xenic morphemes were imported from syllables in other
languages. To parse the entries in unihan, we need something specialized
for understanding these sino-xenic morphemes.