Implement roomaji to hiragana conversion
I'm not entirely sure now, but my thought here was that it would be
easier to work with the Japanese orthography rather than the
romanization; for some reason, Unihan comes with Hangeul but not with
Kana. Hepburn is easy for English speakers to read, but is a bit harder
to process for phonetic components. I may just end up converting back to
nihonsiki.

I thought that there *must* be some existing python package for this,
but there doesn't seem to be anything that is well-maintained. jaconv
and romkan both have very serious errors, with open issues and PR's
ignored at the moment as far as I can tell. For example:

* https://github.com/ikegami-yukino/jaconv/pull/15
* https://github.com/soimort/python-romkan/issues/12

To be fair, the issue I opened with romkan would probably not affect
anything in the Unihan db; however, the lack of maintenance makes me
wary to use it. Additionally, these projects do not seem to focus on
correctness for any particular romanization scheme, but rather handle
the myriad schemes found in natural text, which is a much harder problem
than what I need to solve here.