Add kana -> alpha (IME) conversion
I made up the name "IME" for this format. It's basically how you type
Japanese on a qwerty keyboard, using the still-legal n-apostrophe intead
of they easier-to-type n-n for ん, since it is more readable (e.g.
"on'yomi" is less ambiguous than "onnyomi", etc.). This is overall very
similar to nihonsiki, but doesn't use any diacritics for long vowels. I
find it to be the easiest format for phonetic analysis.