Add on'yomi parsing for Unihan DB
The intent of the parsing is to serve as features for learning decision
trees for pronunciation between languages.

Also add the parses to the unihan expansion step in build-db. This
helped catch many oversights, including some in the romanization phases
(unihan mostly uses hepburn romanization, but still uses "hu" instead
of "fu", for example).

Add a log file for the build-db step, since there were still many
on'yomi which were both simply too esoteric and irregular to warrant
spending time on parsing them. Here's the list of failed ones:

* 仾/HIKUI
* 倜/CHYU
* 僦/SHYUU
* 咶/SHIJI
* 啐/KATSUKACHI
* 嘐/RAUROU
* 圕/TOSHOKAN
* 垔/SHYOU
* 埱/SHYUU
* 塤/TAWAN
* 嬼/RIWA
* 孇/SAWA
* 宨/TEWA
* 寊/CHAWA
* 寠/ROWA
* 帤/GEYO
* 懤/CHYUU
* 掚/KAZARU
* 牃/OUCHOU
* 蒦/KAKUYAKU
* 蔇/KIKE
* 襾/AKA
* 襾/EKE
* 譶/CHYUU
* 醔/SHYUU
* 鈨/HABAKI
* 鏅/SHYUU
* 鏉/SHYUU
* 鐯/JIIYAKU
* 隥/SAKA
* 鬫/HOERU
* 魸/NAMAZU
* 鮱/BORA
* 鮾/KUSARU
* 鰙/HAYA

Some of these represent obviously irregular spellings ("SHYUU", etc.),
and others are just puzzling. They are rare enough that it would take
some time to determine whether these are Unihan errors or just real but
weird readings, such 圕 "toshokan", which is clearly a one-off.

There are some, such as 偟 "kuwau" that I handled but probably should
not have; this appears to be a historical spelling, which in general
is not included Unihan data and therefore should be removed (or marked
as historical somehow in Unihan).