Update configs as learned during work this week
Changes made are:

* Until VSCode can handle poetry's environments better, configure poetry
to put the venv in the same directory.
* Add the VSCode settings so that anyone could just open the project
and start working after installing the dependencies.
* Add comments about the flake8 and black settings being specifically
tailored to work together.
* Coverage doesn't take that long, and it's nicer to have a single
command that runs everything, so just run it on the commit hook. Also,
note that we can specify minimum coverage in the config file.
* Specify the tab size to be 4 spaces.
* Update the readme
* Note that we cannot set the test paths right now due to a segmentation
fault!
* add `setuptools` to the build system dependencies so that the repo
can be installed as-is using just pip. This i