Update dependencies and add programmatic version
For whatever reason, during development (like when running `poetry
run build-db` or using the `poetry install`d version), importlib always
fails to find the uniunihan_db module, immediately throwing an error.
So we have two try statements in __init__.py: one for older pythons and
another for the dev environment.

Also update to python 3.9 so we can use the '|' syntax for combining
dictionaries.