Add HSK vocab
Each level of vocab are in their own file. I did this instead of adding
a python script to download the data because 1) the data has not been
updated since 2012, 2) the size is relatively small and 3) it is only
available for download as an XSL file (not even XSLX!), and I had to
convert it into a useable format using Forvo.