Remove unencoded description characters
These are not very useful as features because they only encode the
number of strokes in the unencoded character. You can't look at them
and recognize the structure of the character, and they are ambiguous
because different DC's with the same number of strokes are represented
using the same character, anyway.