Only keep IDS features with predictive power
Provided the class to be predicted, build a dictionary of the frequency
of unique class values that can be associated with an IDS feature, and
remove any IDS features that aren't associated with at least two
identical values for the class. These features have no predictive power,
and slow down model building. This dropped the number of features from
700 to 450.