Add script to find pronunciation regularities
Experiments with machine learning algorithms made it clear that a
bespoke search of some kind would be the only viable route. This basic
script just looks for groups of characters that share a component and
a pronunciation. There's a lot more that needs to be done, but it's a
good start.

Current score: 1256 characters with no regularities found. We can
improve that!