Find more regularities
This is one step forward and 2 steps back for the moment. We get all
pronunciations for all components and characters. This is a step back
because the complexity of what we've learned is much higher. We'll have
to implement a filter phase of some kind later.

Characters without a regularity: 552
Number of pronunciation groups learned: 1048