Limit to joyo-specific readings
There are many obscure or less-common readings present in Unihan that
are not specified in the joyo kanji table from the Japan ministry of
education. I believe that means that, generally speaking, these readings
would be given in ruby text when encountered by the learner.

Download the joyo kanji table from Wikipedia, and rework it into a CSV
using https://wikitable2csv.ggor.de/ (Wikipedia page table to CSV
converter) and the Japanese regex so helpfully provided here:
https://gist.github.com/terrancesnyder/1345094. Ignore the readings in
parentheses for now, as they are considered secondary. It may be helpful
to put them back to find more regularities.

514 pronunciation groups
78 characters with no pronunciations
952 characters with no regularities

I think this is overall a win. There were 70 characters for which Unihan
had a pronunciation that wasn't standard joyo. There are many fewer
regularities now, but they are more likely to be useful.

Hmm, Joyo pronunciations have only mei for 迷, not mai as in 迷子. That's
kind of worrying, since 迷子 is a very common word.