Ignore 口, since it's in everything
In my mind, it actually could signal ク or コウ, but it's in everything
so we need to be more picky about using it. Even the official joyo data
uses it all over the place, so there's way too much noise in the data.

Found 422 candidate pattern components
Found 606 candidate pronunciation patterns
842 characters with no regularities: