Further weed out non-pure groups
Use the full index of components to pronunciations instead of the
filtered one; this gets rid of the groups that have exceptional
pronunciations.

We've also added some new structures to the index and one warning for
pure groups, which alerts us when there's a character with no
pronunciation with the component of the group. It was going to be more
useful than that, but switching to the non-filtered comp-pron-char index
took care of the other issues that were being flagged.