Redo regularity-finding code
Extract all regularity groups, consider all the pronunciations for all
characters for that component, and then classify that group as pure,
semi-pure, etc. using the rules from Heisig Volume II (number of
readings, number of exceptional characters, etc.).

It's interesting to see where Heisig finagled stuff for memorization;
for example, 方 pattern has way more pronunciations than his because he
put 施 with 也 and 旅 with "readings from everyday words", etc., then
classified the remaining group as much more regular than it really is.
I guess that's fine, since the common words chapter came first and the
learner should only need to learn the remaining group. Although he says
"the material is presented in such a way that... it may be rearranged
freely to suit the particular student's needs." To me this implies that
the chapters could be read out of order; I think the actual case is that
the chapters must be read in the given order, but the groups within the
the chapters can be re-arranged.

Found 391 candidate pattern components
Found 442 candidate pronunciation patterns
78 characters with no pronunciations
997 characters with no regularities
54 characters with unique readings
391 potential groups