Add CKIP Mandarin high frequency word list
I was going to download it from the Wiktionary page where it was shared
originally, but the formatting turned out to be really annoying. I
thought I could use Pandas.read_html, but the first page of words
contains nested tables, and Pandas treats them as new rows, which messes
everything up!

It's good to bring this list back to life, anyway. The original page
where it was hosted is now only available via the wayback machine:

https://web.archive.org/web/20120614131900/http://childes.psy.cmu.edu/morgrams/chinese-xls.zip