Read Hong Kong educational chars with readings
The algorithm for getting the readings may be outdated. I have this old
piece of the FAQ from another package I was inspecting
(https://hackage.haskell.org/package/cjk-0.1.0.1/docs/src/CJK-Data-Unihan-Readings.html):

    This algorithm is based on the Unihan FAQ
    <http://www.unicode.org/faq/han_cjk.html>, which states that the
    best way is to use the kHanyuPinlu, kXHC1983, and kHanyuPinyin
    fields in that order. The kMandarin field may have some readings the
    other three do not but should be used with caution. The kHanyuPinlu
    field lists the most common readings for ideographs in order of
    frequency of use and is the most useful for most purposes. The
    kXHC1983 field contains the most important readings for characters
    in modern use, and the kHanyuPinyin field contains an exhaustive set
    of readings for a large set of characters, but includes obscure
    readings of historic interest only

The current version of the FAQ just says to use the `kMandarin` field if
no context is available for determining the reading. It also states that
the `kMandarin` field is derived algorithmically from the other fields,
but I found that for some characters it contained readings while the
other fields did not. I guess it's possible that experts filled these
missing readings in, but it seems strange to me because these are
supposed to be common hanzi, right? Unless it actually contains some
Cantonese-only characters that experts were needed to provide Mandarin
pronunciations for...