Add historical on'yomi from biglobe
This data comes from http://www5a.biglobe.ne.jp/accent/kana/jion.htm.
I was hoping to generate a CSV programmatically, but this turned out to
be pretty difficult because the page is formatted as nested HTML tables
with no semantic markup. Many of the characters were only provided as
pictures, and these needed to be entered manually. As they all seemed
to be
[CJK compatibility ideographs](https://en.wiktionary.org/wiki/Appendix:Unicode/CJK_Compatibility_Ideographs),
I also added in other compatibility ideographs where they were readily
obvious on Wikipedia (響憎憎漢海謁僧贈廊). I presume that the page was written
before these characters where available in Unicode and has not been
updated since.

I am optimistic that this data will be useful for giving the (very
enthusiastic) learner more context for learning data across different
languages. For example, the Unihan documentation mentions that 和 has
as it's phonetic component 禾. Given that the on'yomi of one is わ
and of the other is か, the relationship is unclear. But given くわ, it's
quite clear that the w was lost in one of the pronunciations over time.
The mandarin is hé and the korean is 화, so the relationship is quite
clear, and the historical on'yomi provides a missing piece of the story
which describes the evolution of the pronunciation.