Use ytenx instead of IDS
This is still WIP because there are many Japanese characters that cannot
be found in ytenx. There are 100 more characters with no regularities
and 100 less that only appear as exceptional, so the numbers make it
look like a wash; however, I'm much more confident in the results, and
and the numbers should improve once we fix the character mismatch.

Warnings:
    116 pronunciation/character combos suggested to move to common words
139 characters with no phonetic component data: {'溫', '芯', ..., '査'}
321 total potential groups:
    144 potential groups of type 1
    83 potential groups of type 2
    11 potential groups of type 3
    46 potential groups of type 4
    37 potential groups of type 5
67 characters with no pronunciations
837 characters with no regularities
268 characters listed only as exceptions
59 characters with unique readings