WIP find correct characters for ytenx lookup
The Joyo characters are often slightly different from those used in
China. Using the current joyo, over 200 cannot be found in ytenx. Using
the old Joyo, about 130 cannot be found. If we allow looking up
kSemanticVariant characters instead, another 30 are found. Using
kZVariant finds another 7. The next solution will be
kCompatibilityVariant, but this field is not contained symmetrically
in Unihan entries, so we will have to create a reverse index.

For example, 頻 and 頻 look exactly the same but were encoded
differently, I guess for historical reasons. One of them has a
compatibility variant link to the other, but the other does not do the
same.

Warnings:
    117 pronunciation/character combos suggested to move to common words
96 characters with no phonetic component data: {'働', '嘆', ... '奪'}
325 total potential groups:
    145 potential groups of type 1
    83 potential groups of type 2
    12 potential groups of type 3
    46 potential groups of type 4
    39 potential groups of type 5
69 characters with no pronunciations
813 characters with no regularities
272 characters listed only as exceptions
57 characters with unique readings