Manually add remaining phonetic components
Add ytenx_ammendment.json, which contains manually-chosen phonetic
characters for Joyo kyuujitai characters and ytenx components. I learned
quite a bit about Unihan doing this, and I think I found important
information that was missing.

With this fix, only characters with no on'yomi have no classification,
and the rest have human-chosen phonetic components, which is trully
wonerful! There's a lot of cleanup that still needs to be done, though.

Warnings:
    117 pronunciation/character combos suggested to move to common words
4 characters with no phonetic component data: {'畑', '込', '峠', '枠'}
340 total potential groups:
    150 potential groups of type 1
    86 potential groups of type 2
    14 potential groups of type 3
    48 potential groups of type 4
    42 potential groups of type 5
74 characters with no pronunciations
739 characters with no regularities
289 characters listed only as exceptions
54 characters with unique readings