Massively simplify group logic and reporting
Now that we don't have to look for groups using IDS, we can get rid of
all of the logic for trying to pick which group a character belongs to.

Also extend the number of purity types, adding SINGLETON and MIXED_D.
MIXED_D has at least 4 characters with at least 2 sharing a
pronunciation. SINGLETON contains a single character.

Remove the "potential" and "candidate" words everywhere, as the groups
are now human-determined and therefore permanent.

Warnings:
    393 pronunciation/character combos suggested to move to common words
4 characters with no phonetic component data: {'峠', '畑', '枠', '込'}
900 total groups:
    151 groups of type 1
    89 groups of type 2
    14 groups of type 3
    51 groups of type 4
    44 groups of type 5
    10 groups of type 6
    121 groups of type 7
    420 groups of type 8
74 characters with no pronunciations
54 characters with unique readings