Find highest frequency words for char/pron pairs
For now we just print the number of char/pron pairings for which we
could successfully find a word, and print the missing char/pron pairs to
a report file.

Note that the number missing is probably a bit inflated right now
because we are using the old joyo glyphs.

4 characters with no phonetic component data: {'枠', '込', '畑', '峠'}
74 characters with no pronunciations
54 characters with unique readings
900 total groups:
    151 PURE groups (390 characters)
    89 SEMI_PURE groups (280 characters)
    14 MIXED_A groups (76 characters)
    51 MIXED_B groups (275 characters)
    44 MIXED_C groups (306 characters)
    10 MIXED_D groups (29 characters)
    121 NO_PATTERN groups (284 characters)
    420 SINGLETON groups (420 characters)
Found words for 1111/2213 words