Handle more cases in Aligner
Add handling for rendaku, renpandaku (I made that word up; just means
ハ→パ), sokuon, and mixed kana/kanji spellings (we always normalize
input to only have Katakana). Add unit tests, too; these helped catch
issues and reduced the complexity overall.

The reason for the `sort` was a little complex. There are two
pronunciations for the character 伴: ハン and バン. But the first can
also undergo rendaku and become the second; this means that if we don't
check for the バン form first, we'll never find the バン pronunciation,
as it will just always be marked as ハン + rendaku.

Found words for 2208/2211 words