Specify more common forms in joyo.csv
Strictly speaking, these aren't the sinjitai forms; they are actually
alternative glyphs. However, the Joyo spec says that they are okay to
use, and the joyo-specified glyphs were almost nowhere to be found in
EDICT, which is a problem!

Wikipedia has this to say about them
(https://en.wikipedia.org/wiki/List_of_j%C5%8Dy%C5%8D_kanji):

>𠮟, 塡, 剝, 頰... are outside of Japan's basic character set, JIS X 0208
(one of them is also outside the Unicode BMP). In practice, these
characters are usually replaced by the characters 叱, 填, 剥, 頬, which
are present in JIS X 0208.

Found words for 2211/2211 words

We got all of them!