Refactor group creation
Simplify the `_index` method logic, moving the index of pronunciations
to characters into `ComponentGroup`.

On accident, I started adding the characters with no pronunciations or
with unique readings to the groups, so the numbers look a bit better.
I actually like this better, as it's more regular, anyway. It doesn't
match the Heisig format as well, and it's possible it may be harder to
memorize... but at this point, for my own personal reference, it's good
like it is.

4 characters with no phonetic component data: {'込', '畑', '峠', '枠'}
74 characters with no pronunciations
54 characters with unique readings
912 total groups:
    165 PURE groups (433 characters)
    89 SEMI_PURE groups (291 characters)
    14 MIXED_A groups (76 characters)
    57 MIXED_B groups (303 characters)
    44 MIXED_C groups (313 characters)
    7 MIXED_D groups (20 characters)
    118 NO_PATTERN groups (282 characters)
    418 SINGLETON groups (418 characters)
Missing words for 0 char/pron pairs
Constructing final output...