Fix pronunciation output
Many pronunciations were being missed in the output, so change the way
they are iterated over, relying on the list of pronunciations stored
with other supplementary character info instead of on the dict of
pronunciations to vocabulary, which still does not have full coverage.

Mark the non-joyo pronunciations and output them in order *after* the
joyo pronunciations.

The 匂/匈 connection is tenuous and it really isn't helpful to put 匂 in
a group with other 凶 characters, so remove it from the the manual
component spec file.

Add more `JP_VOCAB_OVERRIDE` where I find them. The コロナ鍋 one is
really a recent pun, but it already seems to be more well-known than the
one and only example of カ pronunciation I could find for 鍋, so I'll
leave it in the top spot. This seems to be the origin (it's hilarious!):
https://cookpad.com/recipe/6494446

Delete the Netflix reading code, as it's not used anymore. Could
probably remove the Netflix corpus itself, too, but I'll leave it for
now just in case.

`main` was too long for flake8's liking, so put the reporting and final
output logic in their own methods.

5 characters with no phonetic component: {'畑', '匂', '込', '峠', '枠'}
Missing vocab for 52 char/pron pairs
60 characters with unique readings
913 total groups:
    137 PURE groups (360 characters)
    93 SEMI_PURE groups (274 characters)
    15 MIXED_A groups (71 characters)
    54 MIXED_B groups (279 characters)
    60 MIXED_C groups (410 characters)
    27 MIXED_D groups (75 characters)
    109 NO_PATTERN groups (253 characters)
    418 SINGLETON groups (418 characters)