Move joyo reading into utils
Add tests and simplify the logic greatly by splitting the entries for 弁
over multiple lines. Coming up with keywords for each separate old glyph
was tricky, but honestly it doesn't look like the original Joyo document
worked that hard for great keywords, either. "Manage" in particular is
already used twice; I wanted something like "take the reins" but I think
it would be too long and wouldn't match the style of the rest of the
document.