Add kun-yomi to book
Rename kun-yomi entry to kun_yomi to make it easier to refer to in all
contexts, including the jinja template. Add the kun-yomi as a 4th column
in the char-data box, sticking to the right side. Extend the meaning column
into this new fourth column, as it can't overlap with kun-yomi vertically,
anyway, and the meaning strings are quite long.