Test get_variants
Move `get_variants` to be with the data methods, and rename `raw_datasets.py`
to reflect that the data is no longer necessarily raw. Move some widely-used
paths into `constants.py` to fix a circular import issue. Add a test for
`get_variants()` using a small sample of `unihan.json`, which is too big to
load during a unit test.