Unify index methods
Convert hiragana to katakana inside of `JpAligner`, which is the only
place where this normalized form is necessary. Get rid of `JpWord`,
and unify the `index_vocab_{jp,zh}` so that there is only one method
and it takes a list of `Word`s as input.

Get rid of the intermediate formatting stage for the edict/frequency
data. This was needless complexity without full test coverage. Update
the test to use the raw format.