Reorganize main methods
Unify the final output compilation step and place it in a new file.
Factor out other methods where possible, as well, so that the main
methods are fairly straightforward to read.

Several lines related to compiling the component groups could not
be unified because with JP we use the Joyo data in several ways:
old forms are used as the canonical character forms and for determining
components, and new forms are used for finding vocabulary. Update the
architectural plan to reflect that vocab extraction needs to be moved
elsewhere.