Handle Korean character variants and explain 畓
In kyoyuk_hanja.csv, some character fields contained semantic variants
separated by a `/`. Move these to a new `variant` field.

One character, 畓, is actually a gukja, so it needs to have its phonetic
component set manually, and we also add a note field so that we can
tell the reader the interesting history of this character.

It seems that, in China, 田 referred to farmland generally. For whatever
reason, in Japan this character took on the meaning of a rice paddy,
and 畑 was created to refer to other types of farmed fields. Korea also
had their own single-syllable name for a rice paddy, but they wrote it
using the characters for Chinese's 2-syllable word, 水田. At some point
they decided to just squish them conveniently into 畓, with the native
pronunciation 논.