Process group data separately
I realized that we hand't integrated the concept of a cluster into the
char data structure passed along by the `group_chars` step. However,
the downstream consumers already had to deal with deep nesting and
annoying iteration logic, so it makes more sense to just undo the deep
structural integration and pass along the char data and the group index
separately. These can be integrated at a later layout step.