vendorize normalize.css
When reloading the book HTML in the browser to see updates, it seems
to be making network calls to get the CSS from the server again. This
prevents the page from displaying immediately when my internet
connection is bad. normalize.css is really small, so just copy it into
the repository. I don't foresee it being updated often, anyway.