Fix output for 畓
I noticed that 畓, the only character I've written a note for, was
missing in the output. This was because build_db.py needed to be
re-run. Move the last remaining useful method from build_db.py to
datasets.py, and delete the rest of the file. `get_phonetic_components`
now reads the raw data instead of a pre-processed file.

Some other fixes I did on the way:

* Remove zh_template, which is old and unused.
* Add italic styling to character notes
* Correctly output character notes
* Add the phonetic component to `YtenxRhyme`