Select vocab from Kengdic
Unfortunately, the data is almost unusable for now. Only 22K entries
have hanja and an English gloss, and many of the hanja columns don't
contain an alignable surface, just a hanja hint. In the end, only 57
vocabulary could be added! T_T

When we've cleaned up Kengdic, we should also add in a separate download
step, because the `datapackage` package doesn't cache files locally!