Reorganize logging with `loguru`
Get rid of the custom file-based reports everywhere in favor of simply
writing more details with `logger.debug()`. The configuration of
python's built-in loggers proved to be really annoying, so update to
loguru and configure with colored stderr INFO and file DEBUG output.
Only call `configure_logging` from `main()` methods, making the
everything compatible for use as a library.