Separate debug files per script
Use a separate debug output file for each script with a main method
so that we can run them all and still see what warnings there are to be
fixed from the scripts that were run earlier.

With kengdic update, there are now only 82 Korean char/pron pairs
without an example word.