Don't consider 5th tone pronunciations unique
Where a vocab item contains a 5th tone pronunciation, remove it as long
as the same pronunciation with a different tone is also attested. This
is based on the assumption that the 5th tone normally occurs as a part
of a word pronunciation due to stress patterns, not as an inherent tone
for a morpheme. This yields fairly improved regularity statistics for
Mandarin (change in character count shown in square brackets):

    114 PURE groups (290 characters) [+67]
    116 SEMI_PURE groups (379 characters) [+59]
    31 MIXED_A groups (182 characters) [+50]
    92 MIXED_B groups (492 characters) [+98]
    286 MIXED_C groups (2674 characters) [-105]
    31 MIXED_D groups (90 characters) [-63]
    134 NO_PATTERN groups (355 characters) [-1]
    362 SINGLETON groups (362 characters) [-4]
    0 NO_PRONUNCIATIONS groups (0 characters)

Only a handful of 5th tone pronunciations were retained due to not
having any alternatives:

    r5 (兒); alternatives: [er2]
    le5 (了); alternatives: [liao3]
    de5 (地); alternatives: [di4]
    jie5 (價); alternatives: [jia4]
    ga5 (價); alternatives: [jia4]
    shi5 (匙); alternatives: [chi2]
    de5 (的); alternatives: [di4, di1, di2]
    zhe5 (著); alternatives: [zhao1, zhao2, zhuo2, zhu4]
    bei5 (臂); alternatives: [bi4]
    me5 (麼); alternatives: [mo2, ma2]
    qu5 (戌); alternatives: [xu1]
    lo5 (咯); alternatives: [luo4, ge1]
    de5 (底); alternatives: [di3]
    shi5 (殖); alternatives: [zhi2]
    ne5 (呢); alternatives: [ni2]
    bo5 (卜); alternatives: [bu3]
    wo5 (喔); alternatives: [o1]
    shang5 (裳); alternatives: [chang2]

Additionally, we tried to supplement the data using pronunciations from
Unihna, but no alternatives were found for the above characters, so
we deleted the code. It's provided below for reference (was in
add_char_prons.py line 80):

    # check if we can find the proper tone for pronunciations
    # where we only have a 5th tone in the vocab data
    # commented because none were found and this wastes time!
    # for p, p_data in prons.items():
    #     if "5" in p:
    #         p_no_tone = p[:-1]
    #         for up in unihan_prons.keys():
    #             if "5" not in up and up[:-1] == p_no_tone:
    #                 p_data.setdefault("unihan_pron_alt", [])
    #                 p_data["unihan_pron_alt"].append(up)
    #                 print(f"Appending alternative {up} for {p}")