Split output into multiple files
As a single 11MB HTML file, the output was slow to load in the browser.
It could take as much as 30 seconds just to load everything, and that's
annoying because it means that the links are completely unresponsive for
this time.

Split into separate pages for each language/part introduction and each
purity group. The largest file is still 2.8MB, but most of the files
are on the order of 100KB.

Change the char and group ID format to contain the purity group number.
Then update the cross-reference link generation logic to use this
information to link to the correct page.