Remove forwarding from loaded CEDICT
If an entry starts with "see ...", then it's just forwarding the reader
to a homonym to read the definition. These are not interesting or useful
in our book, so we can remove them. Regularity improved very slightly,
which is actually potentially worrying; however, if the dictionary is
forwarding the user to another entry, presumably the other entry is the
more common word and we don't need to worry too much about losing
pronunciations.

284 character(s) with unique readings [+4]
    114 PURE groups (290 characters)
    116 SEMI_PURE groups (379 characters)
    31 MIXED_A groups (182 characters)
    93 MIXED_B groups (499 characters) [+7]
    285 MIXED_C groups (2667 characters) [-7]
    31 MIXED_D groups (90 characters)
    134 NO_PATTERN groups (355 characters)
    362 SINGLETON groups (362 characters)
    0 NO_PRONUNCIATIONS groups (0 characters)