Replace mypy with Pyright
mypy is super slow compared to Pyright, replace usages of mypy with
PyRight in VSCode and in the build verification step.

Use the "basic" setting for PyRight (which is stricter than what we had
for mypy), and fix all type issues (suppress a couple where we are
working with plain dicts).

Also fix line length issues flagged by the changes to .flake8.